requirements.txt
openai
langchain_community
streamlit
python-dotenv
PyPDF2
langchain_openai
langchain

-e .